Se arregla la vista para que sea mas responsive y se modulariza el proyecto para que se tenga mejor orden
