+++ frontend/src/App.vue
@@ -4,9 +4,9 @@
     
     <VertSidebar/>
   
-    <NavBar/>
+    <!-- <NavBar/> -->
     <ProductList/>
-
+    <CarruselOfertas/>
     <FooterBar/>
   </div>
 </template>
@@ -17,6 +17,7 @@ import HeaderBar from './components/HeaderBar.vue';
 import ProductList from './components/ProductList.vue';
 import VertSidebar from './components/VertSidebar.vue';
 import FooterBar from './components/FooterBar.vue';
+import CarruselOfertas from './components/CarruselOfertas.vue';
 
   export default {
     components: {
@@ -25,6 +26,7 @@ import FooterBar from './components/FooterBar.vue';
       ProductList,
       VertSidebar,
       FooterBar,
+      CarruselOfertas
     },
   };  
 </script>
